release: 3.6.4 (#689)
## [v3.6.4] (July 19 2023)
### Feat:
* Create a separate package.json for CommonJS (cjs) module during build
time. This package.json is located under dist/cjs directory. (#687)
* Add a new prop `isUserIdUsedForNickname` to the public interface. This
prop allows using the userId as the nickname. (#683)
* Add an option to the ChannelProvider: `reconnectOnIdle`(default:
true), which prevents data refresh in the background. (#690)

### Fixes:
* Fix an issue where the server returns 32 messages even when requesting
31 messages in the Channel. Now, hasMorePrev will not be set to false
when the result size is larger than the query. (#688)
* Verify the fetched message list size with the requested size of the
MessageListParams. Added a test case for verifying the fetched message
list size. (#686)
* Address the incorrect cjs path in package.json. The common js module
path in the pacakge.json has been fixed. (#685)