release: 3.6.6 (#699)
## [v3.6.6] (Aug 3 2023)
### Feat:
* Add `customExtensionParams` for `sdk.addSendbirdExtensions` (#698)
The 3rd parameter customData to the `sdk.addSendbirdExtension` function,
allowing it to be delivered from outside of UIKit React.
  e.g.
  ```
  <SendbirdProvider
    customExtensionParams={{
a: 'a', // the key-value set will be passed when
sdk.addSendbirdExtensions is called
    }}
  />
  ```
* Call `sdk.addSendbirdExtensions` during the connection process (#682)

### Fixes:
* Change the MessageInput cursor style from disabled to not-allowed;
Thanks @roderickhsiao (#697)
* PendingMsg is missing isUserMessage method (#695)
This resolves the issue where spreading the message object in the
reducer loses some methods like `isUserMessage` and `isFileMessage`
* fix util functions logic of verifying message type. We updated logic
in util functions to verify the message type. (#700)

### Chore:
* Update Trunk-Based Development to Scaled Trunk-Based Development
(#696)
It describes the flow with short-lived feature branches, code review,
and build automation before integrating into main.