feat: add suggested replies (#825)
Addresses https://sendbird.atlassian.net/browse/AC-581 
The feature request is from AI chat bot. 

#### Scenarios 
If a trigger message(can be set from the dashboard) is sent to the
channel from a user, the predefined suggested reply options will be displayed.
Once the user clicks an option from the list, the option list will be gone and the message will be sent.