fix: add workaround to reset IME in mobile webkit (#1108)
## Description
contentEditable does not work as expected in mobile WebKit (Safari).
Events and properties related to composing, necessary for combining
characters like Hangul, do not seem to be passed to IME properly.

When resetting the input with innerHTML, the previous context should be
cleared.
However, if the last text of the previous input was a composable string,
it appears in the next input.

1. If the last string is not a composable string, it is not affected.
This is why preventing the `e.preventDefault()` call when sending a
message with Enter prevents the bug, as the Enter (newline) string is
the last text in the IME context and is not affected by this bug.
2. The bug can be avoided by resetting the IME context(blur or moving
caret) of the last string. This is similar to how moving the cursor
while typing Hangul breaks the composition.

## Changes
In this PR, an attempt is made to reset by moving keyboard focus to
another input, and it operates as follows:

1. A ghost input, which is not visible in the UI but selectable via
focus, is created.
2. After sending the text and resetting with innerHTML, focus
immediately shifts to the ghost input:
   - To prevent the iPhone's virtual keyboard from closing:
- Changing keyboard focus causes the loss of context for the composing
IME string.
- Using `requestAnimationFrame` to move focus back to the original
input:
- To prevent the change of focus from being treated as input → ghost
input → input too quickly, without allowing a minimum interval.
- Using `setTimeout` or similar causes delays due to waiting for the
event loop, resulting in the phenomenon of the virtual keyboard going up
and down. Therefore, `requestAnimationFrame`, associated with the
browser's cycle of drawing new frames, is used.
3. Focus is then returned to the original input.

## Restrictions
The above approach only applies to input from the `Virtual Keyboard`
(like an iPhone keyboard).
The issue still persists when typing using a `Keyboard` in the mobile
device.

ticket: [CLNP-3470]

[CLNP-3470]:
https://sendbird.atlassian.net/browse/CLNP-3470?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ