[CLNP: 2977] fix: m4a format audio file playing in Safari (#1123)
Fixes https://sendbird.atlassian.net/browse/CLNP-2997

#### Problem Statement

Voice audio files created on platforms other than the web, such as
Android/iOS, are in M4A format. When attempting to convert these M4A
files to `audio/mp3` or `audio/mpeg` format for playback on the web,
Safari browser cannot find the correct codec and throws an error:
`Unhandled Promise Rejection: NotSupportedError: The operation is not
supported.`
This results in the file not being playable.

#### Solution

For Safari browser only, if the message file type is M4A, ensure it is
played in this format. However, the file must be parsed as `audio/x-m4a`
instead of `audio/m4a` to be correctly played in the Safari browser.

#### Test Cases Checked

| Browser | .m4a file(created in iOS / Android UIKit) | .mp3
File(created in React UIKit) |
|----------|----------|----------|
| Safari   | ✅       | ✅       |
| Chrome   | ✅       | ✅       |
| Firefox  | ✅       | ✅       |

> A testable account credential for .m4a can be found in the linked Jira
ticket