chore: add checklists to README (#1145)
Addresses https://sendbird.atlassian.net/browse/CLNP-3856 


Added below to the README 
```
### Checklist

Put an `x` in the boxes that apply. You can also fill these out after creating the PR. If unsure, ask the members.
This is a reminder of what we look for before merging your code.

- [ ] **All tests pass locally with my changes**
- [ ] **I have added tests that prove my fix is effective or that my feature works**
- [ ] **Public components / utils / props are appropriately exported**
- [ ] I have added necessary documentation (if appropriate)

```
and cleaned up a bit for the better readability.