[CLNP-5046] Migrate ThreadProvider to the new state management pattern (#1250)
### Overview
This PR migrate ThreadProvider and related files to the new state
management pattern.

### Changelog
* `ThreadProvider` is migrated, and `useThread()` hook is introduced.
* Removed `ThreadDispatcher` usages in ThreadProvider; it is replaced
with the new state-action pattern of `useThread()`.
* `PubSub` of `config` still remains. It is out of scope of this PR.

### Remaining tasks
* Add unit tests and integration tests.

### FurtherConcern
* Handling hook
* The previous `ThreadProvider` contained several custom hooks. Those
hooks retrieved state and actions through `useThreadContext()`
* Due to that, replacing `useThreadContext()` to new `useThread()` faced
a problem. Those hooks conatin `useThread()`, `useThread()` contains
the hooks. So it makes cycle.
* For now, I moved all functionality of the hooks to the `useThread()`,
but it looks wrong. Any good way to handle this?