[CLNP-7306]Fixed a bug that caused a runtime exception when leaving or deleting a channel (#1355)
[fix]: Fixed a bug that caused a runtime exception when leaving or
deleting a channel

Fixes [CLNP-7306](https://sendbird.atlassian.net/browse/CLNP-7306)

### Changelogs
- Fixed a bug that caused a runtime exception when leaving or deleting a
channel

### Checklist

Put an `x` in the boxes that apply. You can also fill these out after
creating the PR. If unsure, ask the members.
This is a reminder of what we look for before merging your code.

- [x] **All tests pass locally with my changes**
- [ ] **I have added tests that prove my fix is effective or that my
feature works**
- [ ] **Public components / utils / props are appropriately exported**
- [ ] I have added necessary documentation (if appropriate)


## External Contributions

This project is not yet set up to accept pull requests from external
contributors.

If you have a pull request that you believe should be accepted, please
contact
the Developer Relations team <developer-advocates@sendbird.com> with
details
and we'll evaluate if we can set up a
[CLA](https://en.wikipedia.org/wiki/Contributor_License_Agreement) to
allow for the contribution.


[CLNP-7306]:
https://sendbird.atlassian.net/browse/CLNP-7306?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ