[CLNP-7360]Multiple Chat Windows cause unexpected behavior (#1358)
[fix]: Multiple Chat Windows cause unexpected behavior

 PR description
- initialState안에 xxxRef들이 Object로 되어 있어서 여러 Provider 인스턴스가 생성될 때 동일한 객체
참조를 공유하는 문제가 있었습니다. 함수로 변경함으로써 각 Provider 인스턴스마다 새로운 초기 상태 객체를 생성하도록 하여
상태 격리 문제를 해결했습니다.

Fixes [CLNP-7360](https://sendbird.atlassian.net/browse/CLNP-7360)

### Changelogs
- Fixed a bug Where Multiple Chat Windows cause unexpected behavior

### Checklist

Put an `x` in the boxes that apply. You can also fill these out after
creating the PR. If unsure, ask the members.
This is a reminder of what we look for before merging your code.

- [x] **All tests pass locally with my changes**
- [ ] **I have added tests that prove my fix is effective or that my
feature works**
- [ ] **Public components / utils / props are appropriately exported**
- [ ] I have added necessary documentation (if appropriate)


## External Contributions

This project is not yet set up to accept pull requests from external
contributors.

If you have a pull request that you believe should be accepted, please
contact
the Developer Relations team <developer-advocates@sendbird.com> with
details
and we'll evaluate if we can set up a
[CLA](https://en.wikipedia.org/wiki/Contributor_License_Agreement) to
allow for the contribution.


[CLNP-7360]:
https://sendbird.atlassian.net/browse/CLNP-7360?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ